批量导入客户信息bug修改

+++ modules/transfer-crm/src/main/java/com/hqjy/mustang/transfer/crm/model/dto/TransferCustomerDTO.java
@@ -143,7 +143,7 @@ public class TransferCustomerDTO {
     /**
      * 工作经验 work_experience
      **/
-    private Byte workExperience;
+    private Integer workExperience;
 
     /**
      * 推广方式：见数据字典GET_WAY枚举(1-主动获取，2-被动获取) get_way
